+++ CHANGES.md
@@ -9,6 +9,7 @@
   the modqueue.
 - Add support for OAuth2 userless (app-only) flow. Both public clients and
   confidential clients are supported. See `Connection.Credentials.t`.
+- Add support for OAuth2 refresh token credentials.
 
 ## Changed